Multiple fixes (description)
> parsecube.h and cube mlx:
Set the width and height of the window to a fixed size.

>game.c:
Fixed int overflow when the value of perp_wall_dist is 0.
Adapt the rest of the ray function to the window size.

>move.c:
user round instead of roundf. Didn't notice improvements with a high movement speed so probably these changes can be redeemed.

>textures.c:
Protected my_mlx_pixel_put function. Probably this can be avoided with the current state of the code (there shouldn't be segfaults here anymore)